+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.10.1] - 2023-11-13
+
+### Fixed
+
+- Fixed bug which caused fatal error in JS if app was opened after local storage was cleared.
+- Fixed MobX console warning.
+
+### Changed
+
+- If no configs exist in the browser's local store, default-created config will be saved back to store even if no changes are made.
+- Removed unneeded console.log messages.
+
 ## [4.10.0] - 2023-11-12
 
 ### Added
@@ -576,7 +588,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added auto-scroll to TX pane, closes #89.
 - Added special delete behaviour for backspace button when in "send on enter" mode, closes #90.
 
-[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v4.10.0...HEAD
+[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v4.10.1...HEAD
+[4.10.1]: https://github.com/gbmhunter/NinjaTerm/compare/v4.9.0...v4.10.1
 [4.10.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.9.0...v4.10.0
 [4.9.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.8.0...v4.9.0
 [4.8.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.7.0...v4.8.0